Extract common delete test case to a new module
To avoid cross importing from different test suites, created a new
test_delete_utils module to hold the common DeleteHostsBaseTestCase.
